+++ app/cmd/cli/query.go
@@ -31,6 +31,8 @@ func init() {
 	queryCmd.AddCommand(queryAppParams)
 	queryCmd.AddCommand(queryNodeReceipts)
 	queryCmd.AddCommand(queryNodeReceipt)
+	queryCmd.AddCommand(queryNodeClaims)
+	queryCmd.AddCommand(queryNodeClaim)
 	queryCmd.AddCommand(queryPocketParams)
 	queryCmd.AddCommand(queryPocketSupportedChains)
 	queryCmd.AddCommand(querySupply)
@@ -566,6 +568,87 @@ var queryAppParams = &cobra.Command{
 	},
 }
 
+var queryNodeClaims = &cobra.Command{
+	Use:   "node-claims <nodeAddr> <height>",
+	Short: "Gets node pending claims for work completed",
+	Long:  `Retrieves the list of all pending proof of work submitted by <nodeAddr> at <height>.`,
+	Args:  cobra.MinimumNArgs(1),
+	Run: func(cmd *cobra.Command, args []string) {
+		app.InitConfig(datadir, tmNode, persistentPeers, seeds, tmRPCPort, tmPeersPort)
+		var height int
+		if len(args) == 1 {
+			height = 0 // latest
+		} else {
+			var err error
+			height, err = strconv.Atoi(args[1])
+			if err != nil {
+				fmt.Println(err)
+				return
+			}
+		}
+		params := rpc.PaginatedHeightAndAddrParams{
+			Height: int64(height),
+			Addr:   args[0],
+		}
+		j, err := json.Marshal(params)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		res, err := QueryRPC(GetNodeClaimsPath, j)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		fmt.Println(res)
+	},
+}
+
+var queryNodeClaim = &cobra.Command{
+	Use:   "node-claim <nodeAddr> <appPubKey> <claimType> <networkId> <sessionHeight> <height>`",
+	Short: "Gets node pending claim for work completed",
+	Long:  `Gets node pending claim for verified proof of work submitted for a specific session`,
+	Args:  cobra.MinimumNArgs(5),
+	Run: func(cmd *cobra.Command, args []string) {
+		app.InitConfig(datadir, tmNode, persistentPeers, seeds, tmRPCPort, tmPeersPort)
+		var height int
+		if len(args) == 5 {
+			height = 0 // latest
+		} else {
+			var err error
+			height, err = strconv.Atoi(args[4])
+			if err != nil {
+				fmt.Println(err)
+				return
+			}
+		}
+		sessionheight, err := strconv.Atoi(args[4])
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		params := rpc.QueryNodeReceiptParam{
+			Address:      args[0],
+			Blockchain:   args[3],
+			AppPubKey:    args[1],
+			SBlockHeight: int64(sessionheight),
+			Height:       int64(height),
+			ReceiptType:  args[2],
+		}
+		j, err := json.Marshal(params)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		res, err := QueryRPC(GetNodeClaimPath, j)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		fmt.Println(res)
+	},
+}
+
 var queryNodeReceipts = &cobra.Command{
 	Use:   "node-receipts <nodeAddr> <height>",
 	Short: "Gets node receipts for work completed",
@@ -584,9 +667,9 @@ var queryNodeReceipts = &cobra.Command{
 				return
 			}
 		}
-		params := rpc.HeightAndAddrParams{
-			Height:  int64(height),
-			Address: args[0],
+		params := rpc.PaginatedHeightAndAddrParams{
+			Height: int64(height),
+			Addr:   args[0],
 		}
 		j, err := json.Marshal(params)
 		if err != nil {